hs-battlegrounds-ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hs-bg-ai"
version = "0.1.0"
description = "Hearthstone Battlegrounds AI Bot"
requires-python = ">=3.10"
dependencies = [
    "watchdog>=3.0",
    "mss>=9.0",
    "Pillow>=10.0",
    "pynput>=1.7",
    "keyboard>=0.13",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "rich>=13.0",
]

[project.optional-dependencies]
ocr = ["paddleocr>=2.7"]
win = ["pywin32>=306"]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
